Backend-Python/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.3
requests==2.32.3
openai==1.30.1
